Move 'set_averagine()' into preprocessing file
