+++ src/python/spectrseqtools/preprocessing/deconvolution.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 """Deconvolution of raw mass spectrometry data."""
 
-import importlib.resources
 from dataclasses import dataclass
 from typing import List, Self, Tuple
 
@@ -137,70 +136,6 @@ class DeconvolutionParameters:
 
         # Return maximum of intensity set by user and found in scan peak set
         return max(min_intensity, min(peak.intensity for peak in scan.peak_set))
-
-
-def set_averagine(backbone: str) -> dict:
-    """
-    Calculate the average elemental composition of RNA.
-
-    Parameters
-    ----------
-    backbone : ["no_backbone", "phosphate", "thiophosphate"]
-        Backbone considered for the composition.
-
-    Returns
-    -------
-    average_composition : dict
-        Dictionary containing average elemental composition.
-
-    Notes
-    -----
-    This function is inspired by https://github.com/koesterlab/oliglow,
-    originally implemented by Moshir Harsh (btemoshir@gmail.com).
-
-    """
-    # Build dict with elemental compositions from file
-    bases = pl.read_csv(
-        importlib.resources.files(__package__)
-        / "../assets"
-        / "elemental_composition.tsv",
-        separator="\t",
-    )
-    base_compositions = [
-        {
-            col: row[bases.get_column_index(col)]
-            for col in bases.columns
-            if col != "base"
-        }
-        for row in bases.iter_rows()
-    ]
-
-    # Calculate average elemental composition
-    average_composition = {}
-    for element in base_compositions[0].keys():
-        average_composition[element] = sum(
-            float(base[element]) for base in base_compositions
-        ) / len(base_compositions)
-
-    # Add backbone elements (if needed)
-    match backbone:
-        case "no_backbone":
-            pass
-        case "phosphate":
-            # Add 1 phosphorus and 2 oxygen for the phosphate group
-            average_composition["O"] += 2
-            average_composition["P"] += 1
-        case "thiophosphate":
-            # Add 1 phosphorus, 1 sulfur, and 1 oxygen for the phosphate group
-            average_composition["O"] += 1
-            average_composition["S"] += 1
-            average_composition["P"] += 1
-        case _:
-            raise NotImplementedError(
-                f"Support for '{backbone}' is currently not given."
-            )
-
-    return average_composition
 
 
 @dataclass
